backend/requirements.txt
fastapi==0.115.6
pydantic-settings==2.13.1
uvicorn[standard]==0.32.1
python-chess==1.999
anthropic==0.40.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.28.1
